+++ test/eex/html_engine_test.exs
@@ -1,0 +1,4 @@
+defmodule EEx.HTMLEngineTest do
+  use ExUnit.Case
+  doctest EEx.HTMLEngine
+end